Moved file tracker to separate file
